requirements.txt
openai>=1.51.0
typer[all]>=0.12.0
python-dotenv>=1.0.0
black>=24.0.0
